fix: code cleanup

+++ feature/core/src/test/java/com/tondracek/myfarmer/core/domain/usecaseresult/DomainResultTest.kt
@@ -81,7 +81,7 @@ class DomainResultTest {
         )
 
         val failure = DomainResult.Failure(AuthError.Unknown).fold(
-            onSuccess = { it as Int },
+            onSuccess = { it },
             onFailure = { -1 }
         )
